+++ packages/cli/src/generator_templates/js/emails/components/BulletedList.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { MjmlText } from "mjml-react";
 import { leadingRelaxed, textBase } from "./theme";