+++ packages/cli/src/generator_templates/ts/emails/Welcome.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Head from "./components/Head";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
@@ -37,13 +36,19 @@ const MyFirstEmail: React.FC<{ name: string }> = ({ name }) => (
       </MjmlSection>
       <MjmlSection padding="0 24px 0" cssClass="smooth">
         <MjmlColumn>
-          <MjmlText padding="0" fontSize={textXl} lineHeight={leadingTight}>
+          <MjmlText
+            cssClass="paragraph"
+            padding="0"
+            fontSize={textXl}
+            lineHeight={leadingTight}
+          >
             Welcome, {name}!
           </MjmlText>
           <MjmlText
             padding="24px 0 16px"
             fontSize={textBase}
             lineHeight={leadingRelaxed}
+            cssClass="paragraph"
           >
             Mailing makes it easy to send great emails from your React app.
             Here’s a bit about Mailing in a nutshell:
@@ -61,11 +66,16 @@ const MyFirstEmail: React.FC<{ name: string }> = ({ name }) => (
           />
           <MjmlSpacer height="24px" />
           <ButtonPrimary
-            link={"https://github.com/successor-software/mailing"}
+            link={"https://github.com/psugihara/mailing"}
             uiText={"View Github Repo"}
           />
           <MjmlSpacer height="24px" />
-          <MjmlText padding="0" fontSize={textBase} lineHeight={leadingRelaxed}>
+          <MjmlText
+            cssClass="paragraph"
+            padding="0"
+            fontSize={textBase}
+            lineHeight={leadingRelaxed}
+          >
             Thank you for checking out our project. We’d love to hear your ideas
             on how we can make sending great emails more fun and simple.
           </MjmlText>
@@ -73,6 +83,7 @@ const MyFirstEmail: React.FC<{ name: string }> = ({ name }) => (
             padding="16px 0 0"
             fontSize={textBase}
             lineHeight={leadingRelaxed}
+            cssClass="paragraph"
           >
             ♥,
             <br />