+++ packages/cli/src/generator_templates/ts/emails/components/ButtonPrimary.tsx
@@ -1,6 +1,5 @@
-import React from "react";
 import { MjmlButton } from "mjml-react";
-import { black } from "./theme";
+import { black, grayLight } from "./theme";
 import { leadingTight, textBase, borderBase } from "./theme";
 
 type ButtonPrimaryProps = {
@@ -10,18 +9,35 @@ type ButtonPrimaryProps = {
 
 const ButtonPrimary: React.FC<ButtonPrimaryProps> = ({ link, uiText }) => {
   return (
-    <MjmlButton
-      lineHeight={leadingTight}
-      fontSize={textBase}
-      height={52}
-      padding="0"
-      align="left"
-      href={link}
-      backgroundColor={black}
-      borderRadius={borderBase}
-    >
-      {uiText}
-    </MjmlButton>
+    <>
+      <MjmlButton
+        lineHeight={leadingTight}
+        fontSize={textBase}
+        height={52}
+        padding="0"
+        align="left"
+        href={link}
+        backgroundColor={black}
+        borderRadius={borderBase}
+        cssClass="light-mode"
+      >
+        {uiText}
+      </MjmlButton>
+      <MjmlButton
+        lineHeight={leadingTight}
+        fontSize={textBase}
+        height={52}
+        padding="0"
+        align="left"
+        href={link}
+        backgroundColor={grayLight}
+        color={black}
+        borderRadius={borderBase}
+        cssClass="dark-mode"
+      >
+        {uiText}
+      </MjmlButton>
+    </>
   );
 };